ryeku-backend/requirements.txt
fastapi~=0.109.2
uvicorn~=0.27.1
python-dotenv~=1.0.1
langchain~=0.1.9
langchain-azure-openai~=0.1.0
langchain-core~=0.1.27
langgraph~=0.0.26
pydantic~=2.6.1
python-multipart~=0.0.9
azure-identity~=1.15.0
azure-keyvault-secrets~=4.7.0
beautifulsoup4~=4.12.3
requests~=2.31.0
aiohttp~=3.9.3
tiktoken~=0.6.0
